@@ -1,7 +1,7 @@
 [tool.black]
 line-length = 88
-target_version = ["py38"]
-exclude = "(.eggs|.git|.hg|.mypy_cache|.nox|.tox|.venv|.svn|_build|buck-out|build|dist|output|data)"
+target_version = ["py310"]
+exclude = "(.eggs|.git|.hg|.mypy_cache|.nox|.tox|.venv|.svn|_build|buck-out|build/|dist/|output/|data/)"
 
 [tool.isort]
 multi_line_output = 3
