@@ -1,14 +1,17 @@
 import gc
+import glob
 import logging
 import os
 import shutil
 import zipfile
 from typing import Callable, List, Optional
 
+import accelerate
 import huggingface_hub
 import numpy as np
 import pandas as pd
 import torch
+import transformers
 import yaml
 from h2o_wave import Q, data, ui
 from jinja2 import Environment, FileSystemLoader
@@ -60,6 +63,10 @@ from llm_studio.src.utils.modeling_utils import load_checkpoint, unwrap_model
 from llm_studio.src.utils.utils import add_file_to_zip, kill_child_processes
 
 logger = logging.getLogger(__name__)
+
+USER_ON_LEFT = False
+USER = USER_ON_LEFT
+BOT = not USER_ON_LEFT
 
 
 async def experiment_start(q: Q) -> None:
@@ -676,7 +683,7 @@ async def experiment_compare(q: Q, selected_rows: list):
             current_name = f" {experiment.name}"
             experiment_names.append(current_name)
 
-        await plot_experiment_charts(q, charts, experiment_names)
+        await charts_tab(q, charts, experiment_names)
 
     elif q.client["experiment/compare/tab"] == "experiment/compare/config":
         if q.client["experiment/compare/diff_toggle"] is None:
@@ -927,159 +934,21 @@ async def experiment_display(q: Q) -> None:
     )
     q.client.delete_cards.add("experiment/display/tab")
 
-    box = ["first", "first", "second", "second"]
-    cnt = 0
-
     if q.client["experiment/display/tab"] == "experiment/display/charts":
-        await plot_experiment_charts(q, [charts], [""])
+        await charts_tab(q, [charts], [""])
     elif q.client["experiment/display/tab"] in [
         "experiment/display/train_data_insights",
         "experiment/display/validation_prediction_insights",
     ]:
-        if (
-            q.client["experiment/display/tab"]
-            == "experiment/display/train_data_insights"
-        ):
-            key = "train_data"
-        elif (
-            q.client["experiment/display/tab"]
-            == "experiment/display/validation_prediction_insights"
-        ):
-            key = "validation_predictions"
-
-        for k1 in ["image", "html"]:
-            if k1 not in charts:
-                continue
-            for k2, v2 in charts[k1].items():
-                if k2 != key:
-                    continue
-                if k1 == "html":
-                    q.page[f"experiment/display/charts/{k1}_{k2}"] = ui.markup_card(
-                        box=box[cnt], title="", content=v2
-                    )
-                    q.client.delete_cards.add(f"experiment/display/charts/{k1}_{k2}")
-
-                    continue
-
-                if k1 == "image":
-                    q.page[f"experiment/display/charts/{k1}_{k2}"] = ui.image_card(
-                        box=box[cnt], title="", type="png", image=v2
-                    )
-                    q.client.delete_cards.add(f"experiment/display/charts/{k1}_{k2}")
-                    continue
+        await insights_tab(charts, q)
     elif q.client["experiment/display/tab"] in ["experiment/display/summary"]:
-        experiment_df = get_experiments(q)
-        experiment_df = experiment_df[experiment_df.id == experiment_id]
-
-        items = []
-
-        cfg = load_config_yaml(
-            os.path.join(q.client["experiment/display/experiment_path"], "cfg.yaml")
-        )
-
-        parent_element = get_parent_element(cfg)
-        if parent_element:
-            items.append(parent_element)
-
-        for col in experiment_df.columns:
-            if col in [
-                "id",
-                "path",
-                "process_id",
-                "status",
-                "eta",
-                "info",
-                "mode",
-                "progress",
-            ]:
-                continue
-            v = experiment_df[col].values[0]
-            if col == "config_file":
-                col = "problem type"
-            t = ui.stat_list_item(label=make_label(col), value=str(v))
-
-            items.append(t)
-
-        q.page["experiment/display/summary"] = ui.stat_list_card(
-            box=box[cnt], items=items, title=""
-        )
-
-        q.client.delete_cards.add("experiment/display/summary")
-
+        await summary_tab(experiment_id, q)
     elif q.client["experiment/display/tab"] in ["experiment/display/config"]:
-        experiment_cfg = load_config_yaml(
-            os.path.join(q.client["experiment/display/experiment_path"], "cfg.yaml")
-        )
-        items = get_cfg_list_items(experiment_cfg)
-
-        q.page["experiment/display/config"] = ui.stat_list_card(
-            box=box[cnt], items=items, title=""
-        )
-
-        q.client.delete_cards.add("experiment/display/config")
+        await configs_tab(q)
     elif q.client["experiment/display/tab"] in ["experiment/display/logs"]:
-        logs_path = f"{q.client['experiment/display/experiment_path']}/logs.log"
-        text = ""
-        in_pre = 0
-        # Read log file only if it already exists
-        if os.path.exists(logs_path):
-            with open(logs_path, "r") as f:
-                for line in f.readlines():
-                    if in_pre == 0:
-                        text += "<div>"
-                    if "INFO: Lock" in line:
-                        continue
-                    # maximum line length
-                    n = 250
-                    chunks = [line[i : i + n] for i in range(0, len(line), n)]
-                    text += "</div><div>".join(chunks)
-
-                    # Check for formatted HTML text
-                    if "<pre>" in line:
-                        in_pre += 1
-                    if "</pre>" in line:
-                        in_pre -= 1
-                    if in_pre == 0:
-                        text += "</div>"
-        items = [ui.text(text)]
-
-        q.page["experiment/display/logs"] = ui.form_card(
-            box=box[cnt], items=items, title=""
-        )
-
-        q.client.delete_cards.add("experiment/display/logs")
-
+        await logs_tab(q)
     elif q.client["experiment/display/tab"] in ["experiment/display/chat"]:
-        cyclic_buffer = data(fields="msg fromUser", size=-500)
-        q.page["experiment/display/chat"] = ui.chatbot_card(
-            box=box[cnt], data=cyclic_buffer, name="experiment/display/chat/chatbot"
-        )
-
-        q.client["experiment/display/chat/box"] = box[cnt]
-        q.client["experiment/display/chat/messages"] = []
-
-        q.client.delete_cards.add("experiment/display/chat")
-
-        message = ["Loading the model...", False]
-        q.page["experiment/display/chat"].data[-1] = message
-
-        await q.page.save()
-
-        logger.info(torch.cuda.memory_allocated())
-
-        cfg, model, tokenizer = load_cfg_model_tokenizer(
-            q.client["experiment/display/experiment_path"]
-        )
-        logger.info(torch.cuda.memory_allocated())
-
-        q.page["experiment/display/chat"].data[-1] = [
-            "Model successfully loaded, how can I help you?",
-            False,
-        ]
-
-        q.client["experiment/display/chat/cfg"] = cfg
-        q.client["experiment/display/chat/model"] = model
-        q.client["experiment/display/chat/tokenizer"] = tokenizer
+        await chat_tab(q)
 
     await q.page.save()
 
@@ -1131,39 +1000,243 @@ async def experiment_display(q: Q) -> None:
     q.client.delete_cards.add("experiment/display/footer")
 
 
-async def parse_param(q: Q, cfg, prompt):
-    prompt = prompt.replace("--", "")
-    parts = prompt.split(" ")
-    args = [" ".join(parts[i : i + 2]) for i in range(0, len(parts), 2)]
-    for arg in args:
-        arg = arg.split(" ")
-        setattr(cfg.prediction, arg[0], type(getattr(cfg.prediction, arg[0]))(arg[1]))
-        q.page["experiment/display/chat"].data[-1] = [
-            f"Permanently changed {arg[0]} to {getattr(cfg.prediction, arg[0])}",
-            False,
+async def insights_tab(charts, q):
+    if q.client["experiment/display/tab"] == "experiment/display/train_data_insights":
+        key = "train_data"
+    elif (
+        q.client["experiment/display/tab"]
+        == "experiment/display/validation_prediction_insights"
+    ):
+        key = "validation_predictions"
+    for k1 in ["image", "html"]:
+        if k1 not in charts:
+            continue
+        for k2, v2 in charts[k1].items():
+            if k2 != key:
+                continue
+            if k1 == "html":
+                q.page[f"experiment/display/charts/{k1}_{k2}"] = ui.markup_card(
+                    box="first", title="", content=v2
+                )
+                q.client.delete_cards.add(f"experiment/display/charts/{k1}_{k2}")
+
+                continue
+
+            if k1 == "image":
+                q.page[f"experiment/display/charts/{k1}_{k2}"] = ui.image_card(
+                    box="first", title="", type="png", image=v2
+                )
+                q.client.delete_cards.add(f"experiment/display/charts/{k1}_{k2}")
+                continue
+
+
+async def summary_tab(experiment_id, q):
+    experiment_df = get_experiments(q)
+    experiment_df = experiment_df[experiment_df.id == experiment_id]
+    items = []
+    cfg = load_config_yaml(
+        os.path.join(q.client["experiment/display/experiment_path"], "cfg.yaml")
+    )
+    parent_element = get_parent_element(cfg)
+    if parent_element:
+        items.append(parent_element)
+    for col in experiment_df.columns:
+        if col in [
+            "id",
+            "path",
+            "process_id",
+            "status",
+            "eta",
+            "info",
+            "mode",
+            "progress",
+        ]:
+            continue
+        v = experiment_df[col].values[0]
+        if col == "config_file":
+            col = "problem type"
+        t = ui.stat_list_item(label=make_label(col), value=str(v))
+
+        items.append(t)
+    q.page["experiment/display/summary"] = ui.stat_list_card(
+        box="first", items=items, title=""
+    )
+    q.client.delete_cards.add("experiment/display/summary")
+
+
+async def configs_tab(q):
+    experiment_cfg = load_config_yaml(
+        os.path.join(q.client["experiment/display/experiment_path"], "cfg.yaml")
+    )
+    items = get_cfg_list_items(experiment_cfg)
+    q.page["experiment/display/config"] = ui.stat_list_card(
+        box="first", items=items, title=""
+    )
+    q.client.delete_cards.add("experiment/display/config")
+
+
+async def logs_tab(q):
+    logs_path = f"{q.client['experiment/display/experiment_path']}/logs.log"
+    text = ""
+    in_pre = 0
+    # Read log file only if it already exists
+    if os.path.exists(logs_path):
+        with open(logs_path, "r") as f:
+            for line in f.readlines():
+                if in_pre == 0:
+                    text += "<div>"
+                if "INFO: Lock" in line:
+                    continue
+                # maximum line length
+                n = 250
+                chunks = [line[i : i + n] for i in range(0, len(line), n)]
+                text += "</div><div>".join(chunks)
+
+                # Check for formatted HTML text
+                if "<pre>" in line:
+                    in_pre += 1
+                if "</pre>" in line:
+                    in_pre -= 1
+                if in_pre == 0:
+                    text += "</div>"
+    items = [ui.text(text)]
+    q.page["experiment/display/logs"] = ui.form_card(box="first", items=items, title="")
+    q.client.delete_cards.add("experiment/display/logs")
+
+
+async def chat_tab(q: Q, load_model=True):
+    running_experiments = get_experiments(q=q)
+    running_experiments = running_experiments[
+        running_experiments.status.isin(["running"])
+    ]
+
+    gpu_blocked = any(
+        [
+            "0" in gpu_list
+            for gpu_list in running_experiments["gpu_list"]
+            .apply(lambda x: x.split(","))
+            .to_list()
         ]
-    return cfg
-
-
-async def experiment_chat(q: Q) -> None:
-    prompt = q.client["experiment/display/chat/chatbot"]
-
-    if prompt.lower().startswith("--"):
-        if prompt.lower() in ["--clean", "--clear"]:
-            q.page["experiment/display/chat"].data[-1] = ["History cleaned", False]
-            q.client["experiment/display/chat/messages"] = []
-            return
-        q.client["experiment/display/chat/cfg"] = await parse_param(
-            q, q.client["experiment/display/chat/cfg"], prompt
+    )
+    if gpu_blocked:
+        q.page["experiment/display/chat"] = ui.form_card(
+            box="first",
+            items=[
+                ui.text(
+                    "Chatbot is not available when GPU "
+                    "is blocked by another experiment."
+                )
+            ],
+            title="",
         )
-        q.client["experiment/display/chat/model"].cfg = q.client[
-            "experiment/display/chat/cfg"
-        ]
+        q.client.delete_cards.add("experiment/display/chat")
         return
 
-    message = [prompt, True]
+    if load_model:
+        loading_message = "Loading the model..."
+    else:
+        loading_message = "Chat History cleaned. How can I help you?"
+
+    q.client.chat_msg_num = "0"
+    cyclic_buffer = data(
+        fields="msg fromUser",
+        size=-500,
+        rows={q.client.chat_msg_num: [loading_message, BOT]},
+    )
+    q.page["experiment/display/chat"] = ui.chatbot_card(
+        box="first", data=cyclic_buffer, name="experiment/display/chat/chatbot"
+    )
+    q.client["experiment/display/chat/messages"] = []
+    q.client.delete_cards.add("experiment/display/chat")
+
+    q.page["experiment/display/chat/settings"] = ui.form_card(
+        box="second",
+        items=[
+            ui.expander(
+                name="chat_settings",
+                label="Chat Settings",
+                items=[ui.progress(label="Loading the model...")],
+                expanded=True,
+            )
+        ],
+    )
+    q.client.delete_cards.add("experiment/display/chat/settings")
+
+    await q.page.save()
+    logger.info(torch.cuda.memory_allocated())
+
+    if load_model:
+        cfg, model, tokenizer = load_cfg_model_tokenizer(
+            q.client["experiment/display/experiment_path"]
+        )
+        q.client["experiment/display/chat/cfg"] = cfg
+        q.client["experiment/display/chat/model"] = model
+        q.client["experiment/display/chat/tokenizer"] = tokenizer
+        initial_message = "Model successfully loaded, how can I help you?"
+
+    else:
+        cfg = q.client["experiment/display/chat/cfg"]
+        assert q.client["experiment/display/chat/model"] is not None
+        assert q.client["experiment/display/chat/tokenizer"] is not None
+        # Message will be replaced, not noticeable in the UI.
+        initial_message = loading_message
+
+    # Hide fields that are should not be visible in the UI
+    cfg.prediction._visibility["metric"] = -1
+    cfg.prediction._visibility["batch_size_inference"] = -1
+    cfg.prediction._visibility["min_length_inference"] = -1
+    cfg.prediction._visibility["stop_tokens"] = -1
+
+    logger.info(torch.cuda.memory_allocated())
+    q.page["experiment/display/chat"].data[q.client.chat_msg_num] = [
+        initial_message,
+        BOT,
+    ]
+
+    option_items = get_ui_elements(
+        cfg=q.client["experiment/display/chat/cfg"].prediction,
+        q=q,
+        pre="chat/cfg_predictions",
+    )
+    q.page["experiment/display/chat/settings"] = ui.form_card(
+        box="second",
+        items=[
+            ui.button(
+                name="experiment/display/chat/clear_history",
+                label="Clear History",
+                primary=True,
+            ),
+            ui.expander(
+                name="chat_settings",
+                label="Chat Settings",
+                items=option_items,
+                expanded=True,
+            ),
+        ],
+    )
+
+
+async def chat_update(q: Q) -> None:
+    """
+    Update the chatbot with the new message.
+    """
+    cfg_prediction = parse_ui_elements(
+        cfg=q.client["experiment/display/chat/cfg"].prediction,
+        q=q,
+        pre="chat/cfg_predictions/cfg/",
+    )
+    logger.info(f"Using chatbot config: {cfg_prediction}")
+    q.client["experiment/display/chat/cfg"].prediction = cfg_prediction
+
+    prompt = q.client["experiment/display/chat/chatbot"]
+
+    message = [prompt, USER]
     q.client["experiment/display/chat/messages"].append(message)
-    q.page["experiment/display/chat"].data[-1] = message
+    q.client.chat_msg_num = q.client.chat_msg_num + "1"
+    q.page["experiment/display/chat"].data[q.client.chat_msg_num] = message
+    q.client.chat_msg_num = q.client.chat_msg_num + "1"
+    q.page["experiment/display/chat"].data[q.client.chat_msg_num] = ["...", BOT]
+    await q.page.save()
 
     cfg = q.client["experiment/display/chat/cfg"]
     model = q.client["experiment/display/chat/model"]
@@ -1174,7 +1247,7 @@ async def experiment_chat(q: Q) -> None:
         for prev_message in q.client["experiment/display/chat/messages"][
             -(cfg.prediction.num_history + 1) :
         ]:
-            if prev_message[1] is True:
+            if prev_message[1] is USER:
                 prev_message = cfg.dataset.dataset_class.parse_prompt(
                     cfg, prev_message[0]
                 )
@@ -1184,7 +1257,7 @@ async def experiment_chat(q: Q) -> None:
                     prev_message += cfg._tokenizer_eos_token
 
             full_prompt += prev_message
-
+    logger.info(f"Full prompt: {full_prompt}")
     inputs = cfg.dataset.dataset_class.encode(
         tokenizer, full_prompt, cfg.tokenizer.max_length_prompt, "left"
     )
@@ -1206,9 +1279,9 @@ async def experiment_chat(q: Q) -> None:
 
     output = cfg.dataset.dataset_class.clean_output(output, [full_prompt], cfg)
 
-    message = [output["predicted_text"][0], False]
+    message = [output["predicted_text"][0], BOT]
     q.client["experiment/display/chat/messages"].append(message)
-    q.page["experiment/display/chat"].data[-1] = message
+    q.page["experiment/display/chat"].data[q.client.chat_msg_num] = message
 
     del output
     del inputs
@@ -1247,7 +1320,7 @@ def unite_validation_metric_charts(charts_list):
     return charts_list
 
 
-async def plot_experiment_charts(q, charts_list, legend_labels):
+async def charts_tab(q, charts_list, legend_labels):
     charts_list = unite_validation_metric_charts(charts_list)
 
     box = ["first", "first", "second", "second"]
@@ -1399,10 +1472,10 @@ async def experiment_artifact_build_error_dialog(q: Q, error: str):
 
 async def experiment_download_artifact(
     q: Q,
-    get_artifact_path_fn: Callable[[str, str, str], str],
-    save_artifact_fn: Callable[[str, str, str, str], str],
-    additional_log: Optional[str],
-    min_disk_space: Optional[float],
+    get_artifact_path_fn: Callable[[str, str], str],
+    save_artifact_fn: Callable[[str, str, str], str],
+    additional_log: Optional[str] = "",
+    min_disk_space: Optional[float] = 0.0,
 ):
     """Download specific artifact, if it does not exist, create it on demand
 
@@ -1524,22 +1597,38 @@ async def experiment_download_model(q: Q, error: str = ""):
 
     if not os.path.exists(zip_path):
         logger.info(f"Creating {zip_path} on demand")
+        cfg = load_config_yaml(os.path.join(experiment_path, "cfg.yaml"))
+
+        device = "cuda"
+        experiments = get_experiments(q)
+        num_running_queued = len(
+            experiments[experiments["status"].isin(["queued", "running"])]
+        )
+        if num_running_queued > 0 or (
+            cfg.training.lora and cfg.architecture.backbone_dtype in ("int4", "int8")
+        ):
+            logger.info("Preparing model on CPU. This might slow down the progress.")
+            device = "cpu"
 
         cfg, model, tokenizer = load_cfg_model_tokenizer(
-            experiment_path, merge=True, device="cpu"
+            experiment_path, merge=True, device=device
         )
 
         model = unwrap_model(model)
         checkpoint_path = cfg.output_directory
-
         model.backbone.save_pretrained(checkpoint_path)
         tokenizer.save_pretrained(checkpoint_path)
 
+        card = get_model_card(cfg, model, repo_id="<path_to_local_folder>")
+        card.save(os.path.join(experiment_path, "model_card.md"))
+
+        logger.info(f"Creating Zip File at {zip_path}")
         zf = zipfile.ZipFile(zip_path, "w")
 
         FILES_TO_PUSH = [
-            "pytorch_model.bin",
             "vocab.json",
+            "sentencepiece.bpe.model",
+            "bpe_encoder.bin",
             "tokenizer_config.json",
             "tokenizer.json",
             "special_tokens_map.json",
@@ -1547,10 +1636,18 @@ async def experiment_download_model(q: Q, error: str = ""):
             "generation_config.json",
             "config.json",
             "added_tokens.json",
+            "model_card.md",
         ]
 
         for file in FILES_TO_PUSH:
-            add_file_to_zip(zf=zf, path=f"{experiment_path}/{file}")
+            path = os.path.join(experiment_path, file)
+            if os.path.isfile(path):
+                add_file_to_zip(zf=zf, path=path)
+
+        # Add model weight files
+        weight_files = glob.glob(os.path.join(checkpoint_path, "pytorch_model*.*"))
+        for file in weight_files:
+            add_file_to_zip(zf=zf, path=file)
 
         zf.close()
 
@@ -1565,14 +1662,54 @@ async def experiment_download_model(q: Q, error: str = ""):
 
 async def experiment_push_to_huggingface_dialog(q: Q, error: str = ""):
     if q.args["experiment/display/push_to_huggingface"] or error:
+        devices = ["cpu"] + [f"cuda:{idx}" for idx in range(torch.cuda.device_count())]
+        default_device = "cuda:0"
+
+        experiments = get_experiments(q)
+        num_running_queued = len(
+            experiments[experiments["status"].isin(["queued", "running"])]
+        )
+        if num_running_queued > 0:
+            default_device = "cpu"
+
+        try:
+            huggingface_hub.login(q.client["default_huggingface_api_token"])
+            user_id = huggingface_hub.whoami()["name"]
+        except Exception:
+            user_id = ""
+
         dialog_items = [
             ui.message_bar("error", error, visible=True if error else False),
             ui.textbox(
+                name="experiment/display/push_to_huggingface/account_name",
+                label="Account Name",
+                value=user_id,
+                width="500px",
+                required=False,
+                tooltip=(
+                    "The account name on HF to push the model to. "
+                    "Leaving it empty will push it to the default user account."
+                ),
+            ),
+            ui.textbox(
                 name="experiment/display/push_to_huggingface/model_name",
                 label="Model Name",
-                value=q.client["experiment/display/experiment"].name,
+                value=q.client["experiment/display/experiment"].name.replace(".", "-"),
                 width="500px",
                 required=True,
+                tooltip="The name of the model as shown on HF.",
+            ),
+            ui.dropdown(
+                name="experiment/display/push_to_huggingface/device",
+                label="Device for preparing the model",
+                required=True,
+                value=default_device,
+                width="500px",
+                choices=[ui.choice(str(d), str(d)) for d in devices],
+                tooltip=(
+                    "The local device to prepare the model before pushing it to HF. "
+                    "CPU can be significantly slower."
+                ),
             ),
             ui.textbox(
                 name="experiment/display/push_to_huggingface/api_key",
@@ -1581,6 +1718,7 @@ async def experiment_push_to_huggingface_dialog(q: Q, error: str = ""):
                 width="500px",
                 password=True,
                 required=True,
+                tooltip="HF API key, needs write access.",
             ),
             ui.buttons(
                 [
@@ -1596,22 +1734,28 @@ async def experiment_push_to_huggingface_dialog(q: Q, error: str = ""):
     elif q.args["experiment/display/push_to_huggingface_submit"]:
         await busy_dialog(
             q=q,
-            title="Exporting to HuggingFace ",
+            title="Exporting to HuggingFace",
             text="Model size can affect the export time significantly.",
         )
 
         experiment_path = q.client["experiment/display/experiment_path"]
         cfg, model, tokenizer = load_cfg_model_tokenizer(
-            experiment_path, merge=True, device="cpu"
+            experiment_path,
+            merge=True,
+            device=q.client["experiment/display/push_to_huggingface/device"],
         )
 
         huggingface_hub.login(
             q.client["experiment/display/push_to_huggingface/api_key"]
         )
-        user_id = huggingface_hub.whoami()["name"]
-        repo_id = (
-            f"{user_id}/{q.client['experiment/display/push_to_huggingface/model_name']}"
-        )
+
+        user_id = q.client["experiment/display/push_to_huggingface/account_name"]
+        if user_id == "":
+            user_id = huggingface_hub.whoami()["name"]
+        exp_name = q.client[
+            "experiment/display/push_to_huggingface/model_name"
+        ].replace(".", "-")
+        repo_id = f"{user_id}/{exp_name}"
 
         # push tokenizer to hub
         tokenizer.push_to_hub(
@@ -1620,30 +1764,7 @@ async def experiment_push_to_huggingface_dialog(q: Q, error: str = ""):
         )
 
         # push model card to hub
-        card_data = huggingface_hub.ModelCardData(
-            language="en",
-            library_name="transformers",
-            tags=["gpt", "llm", "large language model", "h2o-llmstudio"],
-        )
-        card = huggingface_hub.ModelCard.from_template(
-            card_data,
-            template_path="model_card_template.md",
-            base_model=cfg.llm_backbone,  # will be replaced in template if it exists
-            repo_id=repo_id,
-            model_architecture=model.backbone.__repr__(),
-            config=cfg.__repr__(),
-            min_new_tokens=cfg.prediction.min_length_inference,
-            max_new_tokens=cfg.prediction.max_length_inference,
-            do_sample=cfg.prediction.do_sample,
-            num_beams=cfg.prediction.num_beams,
-            temperature=cfg.prediction.temperature,
-            repetition_penalty=cfg.prediction.repetition_penalty,
-            text_prompt_start=cfg.dataset.text_prompt_start,
-            text_answer_separator=cfg.dataset.text_answer_separator,
-            end_of_sentence=cfg._tokenizer_eos_token
-            if cfg.dataset.add_eos_token_to_prompt
-            else "",
-        )
+        card = get_model_card(cfg, model, repo_id)
         card.push_to_hub(
             repo_id=repo_id,
             repo_type="model",
@@ -1722,21 +1843,57 @@ async def experiment_push_to_huggingface_dialog(q: Q, error: str = ""):
     q.client["keep_meta"] = True
 
 
+def get_model_card(cfg, model, repo_id) -> huggingface_hub.ModelCard:
+    card_data = huggingface_hub.ModelCardData(
+        language="en",
+        library_name="transformers",
+        tags=["gpt", "llm", "large language model", "h2o-llmstudio"],
+    )
+    card = huggingface_hub.ModelCard.from_template(
+        card_data,
+        template_path="model_card_template.md",
+        base_model=cfg.llm_backbone,  # will be replaced in template if it exists
+        repo_id=repo_id,
+        model_architecture=model.backbone.__repr__(),
+        config=cfg.__repr__(),
+        use_fast=cfg.tokenizer.use_fast,
+        min_new_tokens=cfg.prediction.min_length_inference,
+        max_new_tokens=cfg.prediction.max_length_inference,
+        do_sample=cfg.prediction.do_sample,
+        num_beams=cfg.prediction.num_beams,
+        temperature=cfg.prediction.temperature,
+        repetition_penalty=cfg.prediction.repetition_penalty,
+        text_prompt_start=cfg.dataset.text_prompt_start,
+        text_answer_separator=cfg.dataset.text_answer_separator,
+        trust_remote_code=cfg.environment.trust_remote_code,
+        transformers_version=transformers.__version__,
+        accelerate_version=accelerate.__version__,
+        torch_version=torch.__version__.split("+")[0],
+        end_of_sentence=cfg._tokenizer_eos_token
+        if cfg.dataset.add_eos_token_to_prompt
+        else "",
+    )
+    return card
+
+
 def load_cfg_model_tokenizer(experiment_path, merge=False, device="cuda:0"):
     cfg = load_config_yaml(os.path.join(experiment_path, "cfg.yaml"))
     cfg.architecture.pretrained = False
     cfg.architecture.gradient_checkpointing = False
     cfg.environment._device = device
     cfg.environment._local_rank = 0
-
-    cfg.prediction.num_history = 2
+    cfg.prediction._visibility["num_history"] = 1
 
     tokenizer = get_tokenizer(cfg)
 
     gc.collect()
     torch.cuda.empty_cache()
 
-    if merge and cfg.training.lora and cfg.architecture.backbone_dtype == "int8":
+    if (
+        merge
+        and cfg.training.lora
+        and cfg.architecture.backbone_dtype in ("int4", "int8")
+    ):
         logger.info("Loading backbone in float16 for merging LORA weights.")
         cfg.architecture.backbone_dtype = "float16"
         cfg.architecture.pretrained = True
@@ -1751,6 +1908,7 @@ def load_cfg_model_tokenizer(experiment_path, merge=False, device="cuda:0"):
         if merge and cfg.training.lora:
             # merges the LoRa layers into the base model.
             # This is needed if one wants to use the base model as a standalone model.
+            logger.info("Merging LORA layers with base model.")
             model.backbone = model.backbone.merge_and_unload()
 
     model = model.to(device).eval()
