@@ -1,6 +1,5 @@
 import codecs
 import collections.abc
-import html
 import logging
 from typing import Any, Dict, List, Tuple, Union
 
@@ -80,7 +79,7 @@ class CustomDataset(Dataset):
         self.prompts = [self.parse_prompt(cfg, prompt) for prompt in self.prompts]
 
         if self.cfg.environment._local_rank == 0:
-            logger.info(f"Sample prompt: {html.escape(self.prompts[0])}")
+            logger.info(f"Sample prompt: {self.prompts[0]}")
 
     @staticmethod
     def parse_prompt(cfg: Any, prompt: str):
@@ -263,23 +262,36 @@ class CustomDataset(Dataset):
         # Read data
         sample = self._read_data(idx=idx, sample=sample)
 
-        # Read labels
-        sample = self._read_label(idx=idx, sample=sample)
-
         return sample
 
-    def _concat_samples(self, prompt, idx):
-        parent = self.prompts[idx] + self.answers[idx]
+    def _get_sample(self, idx):
+        prompt = self.prompts[idx]
+        answer = self.answers[idx]
+
+        prompt_encodings = self.encode(
+            self.tokenizer, prompt, self.cfg.tokenizer.max_length_prompt, "left"
+        )["input_ids"]
         if self.cfg.dataset.add_eos_token_to_answer:
-            parent += self.cfg._tokenizer_eos_token
-        prompt = parent + prompt
-        return prompt
+            max_length_answer = self.cfg.tokenizer.max_length_answer - 1
+        else:
+            max_length_answer = self.cfg.tokenizer.max_length_answer
+        answer_encodings = self.encode(
+            self.tokenizer, answer, max_length_answer, "right"
+        )["input_ids"]
+        if self.cfg.dataset.add_eos_token_to_answer:
+            answer_encodings = torch.cat(
+                [
+                    answer_encodings,
+                    torch.Tensor([self.tokenizer.eos_token_id]),
+                ],
+                dim=0,
+            )
+        return [prompt_encodings, answer_encodings]
 
     def _read_data(self, idx: int, sample: Dict) -> Dict:
         """Reads a single text observation."""
 
-        prompt = self.prompts[idx]
-        answer = self.answers[idx]
+        samples = [self._get_sample(idx)]
 
         if self.parent_ids is not None:
             parent_idx = idx
@@ -292,52 +304,37 @@ class CustomDataset(Dataset):
                     < self.cfg.augmentation.skip_parent_probability
                 ):
                     break
-                prompt = self._concat_samples(prompt, int(parent_idx))
+                samples.insert(0, self._get_sample(int(parent_idx)))
 
         if (
             self.mode == "train"
             and np.random.random() < self.cfg.augmentation.random_parent_probability
         ):
             rnd_idx = np.random.randint(len(self))
-            prompt = self._concat_samples(prompt, rnd_idx)
+            samples.insert(0, self._get_sample(int(rnd_idx)))
 
-        prompt_encodings = self.encode(
-            self.tokenizer, prompt, self.cfg.tokenizer.max_length_prompt, "left"
-        )
+        input_ids = torch.cat([torch.cat(sample) for sample in samples])
+        prompt_mask = torch.cat(
+            [
+                torch.cat([torch.ones_like(sample[0]), torch.zeros_like(sample[1])])
+                for sample in samples
+            ]
+        ).to(torch.bool)
+        attention_mask = torch.ones_like(input_ids)
 
+        labels = input_ids.clone()
+
+        if self.cfg.dataset.mask_prompt_labels:
+            labels.masked_fill_(prompt_mask, -100)
         if self.cfg.dataset.add_eos_token_to_answer:
-            answer_encodings = self.encode(
-                self.tokenizer,
-                answer,
-                self.cfg.tokenizer.max_length_answer - 1,
-                "right",
-            )
-            answer_encodings["input_ids"] = torch.cat(
-                [
-                    answer_encodings["input_ids"],
-                    torch.Tensor([self.tokenizer.eos_token_id]),
-                ],
-                dim=0,
-            )
-            answer_encodings["attention_mask"] = torch.cat(
-                [answer_encodings["attention_mask"], torch.Tensor([1])], dim=0
-            )
-        else:
-            answer_encodings = self.encode(
-                self.tokenizer, answer, self.cfg.tokenizer.max_length_answer, "right"
-            )
-
-        input_ids = torch.cat(
-            [prompt_encodings["input_ids"], answer_encodings["input_ids"]], dim=0
-        )
-        attention_mask = torch.cat(
-            [prompt_encodings["attention_mask"], answer_encodings["attention_mask"]],
-            dim=0,
-        )
+            # eos_token may be equal to pad_token. Add the label back manually.
+            labels[-1] = self.tokenizer.eos_token_id
 
         if self.cfg.tokenizer.max_length < len(input_ids):
-            input_ids = input_ids[-self.cfg.tokenizer.max_length :]
-            attention_mask = attention_mask[-self.cfg.tokenizer.max_length :]
+            labels = labels[-self.cfg.tokenizer.max_length :]
+
+        sample["labels"] = torch.full((self.cfg.tokenizer.max_length,), -100)
+        sample["labels"][-len(labels) :] = labels
 
         sample.update(
             self.pad_tokens(
@@ -348,23 +345,17 @@ class CustomDataset(Dataset):
             )
         )
 
-        sample.update(
-            self.pad_tokens(
-                prompt_encodings["input_ids"],
-                prompt_encodings["attention_mask"],
-                self.cfg.tokenizer.max_length_prompt,
-                self.tokenizer.pad_token_id,
-                prefix="prompt_",
-            )
-        )
+        samples[-1][1] = torch.empty(0)
+        prompt_input_ids = torch.cat([torch.cat(sample) for sample in samples])
+        prompt_attention_mask = torch.ones_like(prompt_input_ids)
 
         sample.update(
             self.pad_tokens(
-                answer_encodings["input_ids"],
-                answer_encodings["attention_mask"],
-                self.cfg.tokenizer.max_length_answer,
+                prompt_input_ids,
+                prompt_attention_mask,
+                self.cfg.tokenizer.max_length_prompt,
                 self.tokenizer.pad_token_id,
-                prefix="answer_",
+                prefix="prompt_",
             )
         )
 
@@ -374,6 +365,11 @@ class CustomDataset(Dataset):
         self, input_ids, attention_mask, max_length, pad_token_id, prefix=""
     ):
         sample = {}
+
+        if max_length < len(input_ids):
+            input_ids = input_ids[-max_length:]
+            attention_mask = attention_mask[-max_length:]
+
         sample[f"{prefix}input_ids"] = torch.full((max_length,), pad_token_id)
         sample[f"{prefix}input_ids"][-len(input_ids) :] = input_ids
         sample[f"{prefix}attention_mask"] = torch.zeros(max_length)
@@ -392,18 +388,3 @@ class CustomDataset(Dataset):
             encodings["input_ids"] = encodings["input_ids"][-max_length:]
             encodings["attention_mask"] = encodings["attention_mask"][-max_length:]
         return encodings
-
-    def _read_label(self, idx, sample) -> Dict:
-        sample["labels"] = sample["input_ids"].clone()
-
-        eos_at_end = self.tokenizer.eos_token_id == sample["input_ids"][-1]
-        sample["labels"][sample["labels"] == self.tokenizer.pad_token_id] = -100
-        if self.cfg.dataset.mask_prompt_labels:
-            sample["labels"][
-                : int(-sample["answer_attention_mask"].sum().item())
-            ] = -100
-        if eos_at_end:
-            # eos_token may be equal to pad_token. Add the label back manually.
-            sample["labels"][-1] = self.tokenizer.eos_token_id
-
-        return sample
